\name{biouml.analysis.list}
\alias{biouml.analysis.list}
\title{List available BioUML analyses}
\description{\code{biouml.analysis.list} fetches list of available analyses from current BioUML server}
\usage{biouml.analysis.list()}
\value{
  A data frame (\code{\link{data.frame}}) with two column 'Group' and 'Name'.
}
