\name{biouml.import}
\alias{biouml.import}
\title{Import file to BioUML}
\description{\code{biouml.import} imports file to BioUML repository}
\usage{biouml.import(file, parentPath, importer, importer.params=list())}
\arguments{
  \item{file}{The name of file to import}
  \item{parentPath}{Path to folder in BioUML repository}
  \item{importer}{character string specifying format}
  \item{importer.params}{list of parameters to exporter}
}
\value{
  Resulting path in BioUML repository
}
