\name{biouml.importers}
\alias{biouml.importers}
\title{List BioUML importers}
\description{\code{biouml.importers} fetches the list of importers from BioUML server, these importers can be used in \code{\link{biouml.import}} function}
\usage{biouml.importers()}
\value{
  Character vector of BioUML importers.
}
